+++ packages/subgraph/tests/token/utils.ts
@@ -13,7 +13,7 @@ import {
   ADDRESS_ONE,
   DAO_ADDRESS,
   PROPOSAL_ENTITY_ID,
-  PROPOSAL_ID,
+  PLUGIN_PROPOSAL_ID,
   CONTRACT_ADDRESS,
   VOTING_MODE,
   SUPPORT_THRESHOLD,
@@ -297,7 +297,7 @@ export function createTokenVotingProposalEntityState(
   dao: string = DAO_ADDRESS,
   pkg: string = CONTRACT_ADDRESS,
   creator: string = ADDRESS_ONE,
-  proposalId: string = PROPOSAL_ID,
+  pluginProposalId: string = PLUGIN_PROPOSAL_ID,
 
   open: boolean = true,
   executed: boolean = false,
@@ -319,7 +319,7 @@ export function createTokenVotingProposalEntityState(
   let tokenVotingProposal = new TokenVotingProposal(entityID);
   tokenVotingProposal.dao = Address.fromString(dao).toHexString();
   tokenVotingProposal.plugin = Address.fromString(pkg).toHexString();
-  tokenVotingProposal.proposalId = BigInt.fromString(proposalId);
+  tokenVotingProposal.pluginProposalId = BigInt.fromString(pluginProposalId);
   tokenVotingProposal.creator = Address.fromString(creator);
 
   tokenVotingProposal.open = open;